Move FormatDuration to internal/ui and drop per-run summary lines
FormatDuration is a display utility; it belongs in internal/ui alongside
ErrError/ErrSuccess/ErrWarning rather than in internal/validate. Callers
in both validate and cmd now import it from ui.

Also removes the "N passed  Xs" aggregate summary from RunAll and RunRemote.
The cmd layer already owns end-of-run timing via its own start timer, so the
per-run summary was a duplicate with a slightly different elapsed value.
